backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
alembic>=1.12.0
python-multipart>=0.0.6

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.24.0
Pillow>=10.0.0
